#4 AI Backend Engineering (1)/day41-45_starter-code/pyproject.toml
# uv 패키지 관리자를 사용하는 프로젝트 설정 파일
#
# 사용법:
#   uv sync          # 의존성 설치
#   uv run uvicorn app.main:app --reload  # 서버 실행

[project]
name = "lumi-agent"
version = "0.3.0"  
description = "루미(Lumi) - 버추얼 아이돌 AI 에이전트 서비스"
readme = "README.md"
requires-python = ">=3.11"

# 핵심 의존성
# - LangGraph: 에이전트 워크플로우 프레임워크
# - FastAPI: 비동기 웹 프레임워크
# - Upstage: Solar LLM API
# - Supabase: PostgreSQL + pgvector 데이터베이스
dependencies = [
    # ===== LangGraph & LangChain =====
    "langgraph>=1.0.0",                  # 그래프 기반 에이전트 프레임워크
    "langchain-core>=1.0.0",             # LangChain 핵심 모듈
    "langchain-upstage>=0.5.0",          # Upstage Solar API 연동
    "langchain-community>=0.3.0",        # 커뮤니티 통합 모듈

    # ===== Web Framework =====
    "fastapi>=0.125.0",                  # 비동기 웹 프레임워크
    "uvicorn[standard]>=0.32.0",         # ASGI 서버

    # ===== Database & Cache =====
    "supabase>=2.20.0",                  # Supabase Python 클라이언트

    # ===== Configuration =====
    "pydantic>=2.0.0",                   # 데이터 검증
    "pydantic-settings>=2.0.0",          # 환경변수 설정 관리
    "python-dotenv>=1.0.0",              # .env 파일 로드

    # ===== Utilities =====
    "httpx>=0.27.0",                     # 비동기 HTTP 클라이언트
    "loguru>=0.7.0",                     # 로깅 라이브러리
    "orjson>=3.10.0",                    # 빠른 JSON 처리

    "gradio>=5.0.0",                     # 웹 UI 프레임워크
    "sse-starlette>=2.1.0",              # Server-Sent Events 지원
]

[project.optional-dependencies]
# 개발용 의존성
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "ruff>=0.8.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.black]
line-length = 88
target-version = ["py311"]
